added single click handler for flags on victory points tracker
